+++ NEWS.md
@@ -15,7 +15,7 @@ v0.2.12
 -------
 ## Bug fixes
 
-- `NetCDFWriter` now correctly writes purely vertical spaces.
+- `NetCDFWriter` now correctly writes purely vertical and point spaces.
 
 v0.2.11
 -------